Move BlogTests out of the Blog model
